ref(ui): Make page time range selector reusable
